+++ tapeagents/core.py
@@ -4,10 +4,12 @@ Core data structures for the tape agents framework.
 
 from __future__ import annotations
 
+import datetime
 import json
-from typing import Any, Generic, Iterable, Iterator, List, Literal, TypeVar
+from typing import Any, Generic, Iterable, Iterator, List, Literal, TypeAlias, TypeVar
 from uuid import uuid4
 
+import litellm
 from pydantic import BaseModel, Field, SerializeAsAny
 from typing_extensions import Self
 
@@ -349,6 +351,34 @@ class Prompt(BaseModel):
 
     def __bool__(self) -> bool:
         return bool(self.messages)
+
+
+LLMOutput: TypeAlias = litellm.utils.Message
+"""Type alias for the output of the language model."""
+
+
+class LLMCall(BaseModel):
+    """
+    LLMCall stores info about a call to a language model.
+
+    Attributes:
+        timestamp (str): The timestamp when the call was made, in ISO 8601 format.
+        prompt (Prompt): The input prompt provided to the language model.
+        output (LLMOutput): The output generated by the language model.
+        prompt_length_tokens (int): The length of the prompt in tokens. Defaults to -1 if not set.
+        output_length_tokens (int): The length of the output in tokens. Defaults to -1 if not set.
+        cached (bool): Indicates whether the result was retrieved from cache.
+    """
+
+    timestamp: str = Field(default_factory=lambda: datetime.datetime.now().isoformat())
+    prompt: Prompt
+    output: LLMOutput
+    prompt_length_tokens: int = -1
+    output_length_tokens: int = -1
+    cached: bool
+    llm_info: dict = {}
+    cost: float = 0
+    logprobs: list[TokenLogprob] = Field(default_factory=list, exclude=True)
 
 
 class TokenLogprob(BaseModel):
