+++ tapeagents/llms/__init__.py
@@ -1,8 +1,9 @@
-from .base import LLM
+from tapeagents.core import LLMCall, LLMOutput
+
+from .base import LLM, LLMEvent, LLMStream
 from .litellm import LiteLLM
 from .mock import MockLLM
 from .replay import ReplayLLM
 from .trainable import TrainableLLM
-from .types import LLMCall, LLMEvent, LLMOutput, LLMStream
 
 __all__ = ["LLM", "LLMCall", "LLMOutput", "LLMStream", "LLMEvent", "LiteLLM", "ReplayLLM", "TrainableLLM", "MockLLM"]
