+++ tapeagents/observe.py
@@ -14,8 +14,7 @@ from typing import Callable, Optional, Type
 from pydantic import BaseModel
 
 from tapeagents.config import sqlite_db_path
-from tapeagents.core import Prompt, Tape
-from tapeagents.llms.types import LLMCall, LLMOutput
+from tapeagents.core import LLMCall, LLMOutput, Prompt, Tape
 
 logger = logging.getLogger(__name__)
